Exam/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
sqlalchemy==2.0.39
alembic==1.15.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
pydantic==2.10.6
email-validator==2.2.0
opencv-python==4.11.0.86
numpy==2.2.4
Pillow==11.1.0
reportlab==4.3.1
python-dotenv==1.0.1
bcrypt==4.0.1
